requirements.txt
fastapi>=0.78.0,<0.79.0
uvicorn>=0.18.2,<0.19.0
sqlalchemy>=1.4.39,<1.5.0
passlib>=1.7.4,<1.8.0
bcrypt>=3.2.2,<3.3.0
python-jose>=3.3.0,<3.4.0
python-multipart>=0.0.5,<0.1.0
black
isort